Removed old JPEG, use current working version

+++ compressibility_scorer.py
@@ -40,19 +40,6 @@ class JPEG_class:
         
         return -np.array(sizes)
     
-
-
-### JPEG Compression
-# def jpeg_compressibility(images):
-#     sizes = []
-#     for pil_img in images:
-#         buffer = io.BytesIO()
-#         pil_img.save(buffer, format="JPEG", quality=95)
-#         size = buffer.tell() / 1000.0
-#         buffer.close()
-#         sizes.append(size)
-#     return -np.array(sizes)
-
 def jpeg_compressibility(images):
     if isinstance(images, torch.Tensor):
         images = (images * 255).round().clamp(0, 255).to(torch.uint8).cpu().numpy()
@@ -77,7 +64,7 @@ def jpeg_compressibility(images):
             image.save(buffer, format="JPEG", quality=95)
             sizes.append(buffer.tell() / 1000)  # Size in kilobytes
     
-    return -np.array(sizes)
+    return -np.array(sizes).astype(np.float32)
 
 def classify_compressibility_scores(y):
     # Applying thresholds to map scores to classes
